% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_terms.R
\name{clean_terms}
\alias{clean_terms}
\title{Clean terms}
\usage{
clean_terms(dictionary = dictionary_path, x)
}
\arguments{
\item{dictionary}{Dictionary}

\item{x}{X}
}
\description{
Clean terms
}
\examples{
NULL
}
