% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_terms.R
\name{clean_terms_2}
\alias{clean_terms_2}
\title{Clean terms 2}
\usage{
clean_terms_2(dictionary = dictionary_path, x)
}
\arguments{
\item{dictionary}{Dictionary}

\item{x}{X}
}
\description{
Clean terms 2
}
\examples{
NULL
}
