% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_ion_taste_intensities.R
\name{correlate_ion_taste_intensities}
\alias{correlate_ion_taste_intensities}
\title{Correlate ion taste intensities}
\usage{
correlate_ion_taste_intensities(
  input_ions = "./data/fractions_mzmine/fractions.csv",
  input_tastes = NULL,
  output = "./data/correlations.tsv",
  min_jury = 2L,
  min_area_ion = 1L,
  imputation_factor = 0.5,
  widths = 5:9
)
}
\arguments{
\item{input_ions}{Input ions}

\item{input_tastes}{Input tastes}

\item{output}{Output}

\item{min_jury}{Min jury}

\item{min_area_ion}{Min area ion}

\item{imputation_factor}{Imputation factor}

\item{widths}{Widths}
}
\description{
Correlate ion taste intensities
}
\examples{
NULL
}
