% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_chasselas.R
\name{export_chasselas}
\alias{export_chasselas}
\title{Export chasselas}
\usage{
export_chasselas(
  path = analysis_path_04_output,
  product_name = PRODUCTNAME,
  cluster = CLUSTER,
  date = DATE
)
}
\arguments{
\item{path}{Path}

\item{product_name}{Product name}

\item{cluster}{Cluster}

\item{date}{Date}
}
\description{
Export chasselas
}
\examples{
NULL
}
