% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sigmoid.R
\name{geom_sigmoid}
\alias{geom_sigmoid}
\title{geom_sigmoid}
\usage{
geom_sigmoid(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  smooth = 8,
  direction = "x",
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{provide you own mapping. both x, xend, y and yend need to be numeric.}

\item{data}{provide you own data}

\item{geom}{xhange geom}

\item{position}{change position}

\item{na.rm}{remove missing values}

\item{show.legend}{show legend in plot}

\item{smooth}{how much smooth should the curve have? More means steeper curve.}

\item{direction}{the character x or y depending of smoothing direction}

\item{inherit.aes}{should the geom inherits aestethics}

\item{...}{other arguments to be passed to the geom}
}
\value{
ggplot layer
}
\description{
geom_sigmoid
}
