% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chasselas_modulation.R
\name{plot_chasselas_modulation}
\alias{plot_chasselas_modulation}
\title{Plot chasselas modulation}
\usage{
plot_chasselas_modulation(
  input = NULL,
  output = "./data/figures/figure_modulation.pdf"
)
}
\arguments{
\item{input}{Input}

\item{output}{Output}
}
\description{
Plot chasselas modulation
}
\examples{
NULL
}
