% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chasselas_variation.R
\name{plot_chasselas_variation}
\alias{plot_chasselas_variation}
\title{Plot chasselas variation}
\usage{
plot_chasselas_variation(
  input = NULL,
  output_jury = "./data/figures/figure_variation_jury.pdf",
  output_session = "./data/figures/figure_variation_session.pdf"
)
}
\arguments{
\item{input}{Input}

\item{output_jury}{Output jury}

\item{output_session}{Output session}
}
\description{
Plot chasselas variation
}
\examples{
NULL
}
