% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chromatograms.R
\name{plot_chromatograms}
\alias{plot_chromatograms}
\title{Plot chromatograms}
\usage{
plot_chromatograms(
  input_dir = "./data/20210619",
  input_groups = NULL,
  raw_index = 7,
  fractions_indices = 12:65,
  xlim = c(2, 25),
  output = "./data/figures/figure_chromatograms.pdf"
)
}
\arguments{
\item{input_dir}{Input dir}

\item{input_groups}{Input groups}

\item{raw_index}{Raw index}

\item{fractions_indices}{Fractions indices}

\item{xlim}{Xlim}

\item{output}{Output}
}
\description{
Plot chromatograms
}
\examples{
NULL
}
