% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concentration.R
\name{plot_concentration}
\alias{plot_concentration}
\title{Plot concentration}
\usage{
plot_concentration(
  input = NULL,
  output = "./data/figures/figure_raw_extract.pdf"
)
}
\arguments{
\item{input}{Input}

\item{output}{Output}
}
\description{
Plot concentration
}
\examples{
NULL
}
