% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{plot_correlations}
\alias{plot_correlations}
\title{Plot correlations}
\usage{
plot_correlations(
  input_correlations = "./data/correlations.tsv",
  input_ions = "./data/fractions_mzmine/fractions.csv",
  input_tastes = NULL,
  output_1 = "./data/figures/figure_correlations_1.pdf",
  output_2 = "./data/figures/figure_correlations_2.pdf",
  output_3 = "./data/figures/figure_correlations_3.pdf",
  annotation_path_fractions =
    "./data/processed/241217_130815_fractions/fractions_results.tsv",
  imputation_factor = 0.5,
  max_pval = 0.05,
  min_area_ion = 1L,
  min_confidence = 0.45,
  min_consistency = 0.9,
  min_corr = 0.95,
  min_intense_ions_ratio = 0.8,
  min_jury = 2L,
  mode = "pos",
  widths = 5:9
)
}
\arguments{
\item{input_correlations}{Input correlations}

\item{input_ions}{Input ions}

\item{input_tastes}{Input tastes}

\item{output_1}{Output 1}

\item{output_2}{Output 2}

\item{output_3}{Output 3}

\item{annotation_path_fractions}{Annotation path fractions}

\item{imputation_factor}{Imputation factor}

\item{max_pval}{Max p value}

\item{min_area_ion}{Minimal area ion}

\item{min_confidence}{Minimal confidence}

\item{min_consistency}{Minimal consistency}

\item{min_corr}{Minimal correlation}

\item{min_intense_ions_ratio}{Minimal intense ions ratio}

\item{min_jury}{Minimal number of jury}

\item{mode}{Mode}

\item{widths}{Widths (of the taste-ion windows)}
}
\description{
Plot correlations
}
\examples{
NULL
}
