% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_descriptors_network.R
\name{plot_descriptors_network}
\alias{plot_descriptors_network}
\title{Plot descriptors network}
\usage{
plot_descriptors_network(
  input = NULL,
  output = "./data/figures/figure_network.pdf"
)
}
\arguments{
\item{input}{Input}

\item{output}{Output}
}
\description{
Plot descriptors network
}
\examples{
NULL

}
