% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_napping.R
\name{plot_napping}
\alias{plot_napping}
\title{Plot napping}
\usage{
plot_napping(
  input_coordinates = system.file("extdata", "napping_coordinates.tsv", package =
    "sapid"),
  input_descriptors = system.file("extdata", "napping_descriptors.tsv", package =
    "sapid"),
  sessions = 2
)
}
\arguments{
\item{input_coordinates}{Input coordinates}

\item{input_descriptors}{Input descriptors}

\item{sessions}{Sessions}
}
\description{
Plot napping
}
\examples{
NULL
}
