% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profiles.R
\name{plot_profiles}
\alias{plot_profiles}
\title{Plot profiles}
\usage{
plot_profiles(
  input = NULL,
  output = "./data/figures/figure_profiles.pdf",
  annotation_path_extract =
    "~/Git/sapid/data/processed/241026_103144_extract/extract_results.tsv",
  annotation_path_fractions =
    "~/Git/sapid/data/processed/241217_130815_fractions/fractions_results.tsv",
  features_path_extract = "~/Git/sapid/data/extract_mzmine/extract.csv",
  features_path_fractions = "~/Git/sapid/data/fractions_mzmine/fractions.csv",
  peaks_dir_extract = "~/Git/sapid/data/interim/peaks/extract",
  peaks_dir_fractions = "~/Git/sapid/data/interim/peaks/fractions",
  min_confidence = 0.45,
  min_similarity_prefilter = 0.6,
  min_similarity_filter = 0.8,
  mode = "pos",
  detector = "ms",
  type = "analysis"
)
}
\arguments{
\item{input}{Input}

\item{output}{Output}

\item{annotation_path_extract}{Annotation path extract}

\item{annotation_path_fractions}{Annotation path fractions}

\item{features_path_extract}{Features path extract}

\item{features_path_fractions}{Annotation path fractions}

\item{peaks_dir_extract}{Peaks dir extract}

\item{peaks_dir_fractions}{Peaks dir fractions}

\item{min_confidence}{Minimal confidence}

\item{min_similarity_prefilter}{Minimal similarity prefilter}

\item{min_similarity_filter}{Minimal similarity filter}

\item{mode}{Mode}

\item{detector}{Detector}

\item{type}{Type}
}
\description{
Plot profiles
}
\examples{
NULL
}
