% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_chasselas.R
\name{prepare_chasselas}
\alias{prepare_chasselas}
\title{Prepare chasselas}
\usage{
prepare_chasselas(
  input_dir = "~/switchdrive/SAPERE/02_raw-data/inhouse/02_sensory",
  sessions = seq(1, 7),
  output = "inst/extdata/chasselas.tsv"
)
}
\arguments{
\item{input_dir}{Input dir}

\item{sessions}{Sessions}

\item{output}{output}
}
\description{
Prepare chasselas
}
\examples{
NULL
}
