% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_napping.R
\name{prepare_napping}
\alias{prepare_napping}
\title{Prepare napping}
\usage{
prepare_napping(
  input_dir = "~/switchdrive/SAPERE/02_raw-data/inhouse/02_sensory",
  dictionary_generic_path = system.file("extdata", "dictionary_generic.tsv", package =
    "sapid"),
  dictionary_napping_path = system.file("extdata", "dictionary_napping.tsv", package =
    "sapid"),
  dictionary_specific_path = system.file("extdata", "dictionary_specific.tsv", package =
    "sapid"),
  sessions = seq(1, 8),
  output_coordinates = "inst/extdata/napping_coordinates.tsv",
  output_descriptors = "inst/extdata/napping_descriptors.tsv"
)
}
\arguments{
\item{input_dir}{Input directory}

\item{dictionary_generic_path}{Dictionary generic path}

\item{dictionary_napping_path}{Dictionary napping path}

\item{dictionary_specific_path}{Dictionary specific path}

\item{sessions}{Sessions}

\item{output_coordinates}{Output coordinates}

\item{output_descriptors}{Output descriptors}
}
\description{
Prepare napping
}
\examples{
NULL
}
