% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_profiles.R
\name{prepare_profiles}
\alias{prepare_profiles}
\title{Prepare profiles}
\usage{
prepare_profiles(
  input_dir = "~/switchdrive/SAPERE/02_raw-data/inhouse/02_sensory",
  dictionary_generic_path = system.file("extdata", "dictionary_generic.tsv", package =
    "sapid"),
  dictionary_specific_path = system.file("extdata", "dictionary_specific.tsv", package =
    "sapid"),
  sessions = seq(1, 7),
  output = "inst/extdata/profiles.tsv"
)
}
\arguments{
\item{input_dir}{Input dir}

\item{dictionary_generic_path}{Dictionary generic path}

\item{dictionary_specific_path}{Dictionary specific path}

\item{sessions}{Sessions}

\item{output}{output}
}
\description{
Prepare profiles
}
\examples{
NULL
}
