% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapid-package.R
\docType{package}
\name{sapid-package}
\alias{sapid}
\alias{sapid-package}
\title{sapid: A Strategy to Analyze Plant Extracts Taste In Depth}
\description{
This package provides the infrastructure to implement a Strategy to Analyze Plant Extracts Taste In Depth.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adafede/sapid}
  \item \url{https://adafede.github.io/sapid}
  \item Report bugs at \url{https://github.com/adafede/sapid/issues}
}

}
\author{
\strong{Maintainer}: Adriano Rutz \email{adafede@gmail.com} (\href{https://orcid.org/0000-0003-0443-9902}{ORCID})

}
\keyword{internal}
