% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_chasselas.R
\name{treat_chasselas}
\alias{treat_chasselas}
\title{Treat chasselas}
\usage{
treat_chasselas(
  path = data_inhouse_sensory_path,
  product_name = PRODUCTNAME,
  cluster = CLUSTER,
  date = DATE
)
}
\arguments{
\item{path}{Path}

\item{product_name}{Product name}

\item{cluster}{Cluster}

\item{date}{Date}
}
\description{
Treat chasselas
}
\examples{
NULL
}
