% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_consistent_profiles.R
\name{load_consistent_profiles}
\alias{load_consistent_profiles}
\title{Load consistent profiles}
\usage{
load_consistent_profiles(input, min_jury = 2L)
}
\arguments{
\item{input}{Input}

\item{min_jury}{Min jury. Default to 2.}
}
\description{
Load consistent profiles
}
\examples{
NULL
}
