% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_session.R
\name{load_session}
\alias{load_session}
\title{Load session}
\usage{
load_session(input_dir, session_info, tab)
}
\arguments{
\item{input_dir}{Input dir}

\item{session_info}{Session info}

\item{tab}{Tab}
}
\description{
Load session
}
\examples{
NULL
}
