% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chasselas_variation.R
\name{plot_chasselas_variation}
\alias{plot_chasselas_variation}
\title{Plot chasselas variation}
\usage{
plot_chasselas_variation(
  input = system.file("extdata", "chasselas.tsv", package = "sapid"),
  output_jury = "~/switchdrive/SAPERE/06_figures/figure_variation_jury.svg",
  output_session = "~/switchdrive/SAPERE/06_figures/figure_variation_session.svg"
)
}
\arguments{
\item{input}{Input}

\item{output_jury}{Output jury}

\item{output_session}{Output session}
}
\description{
Plot chasselas variation
}
\examples{
NULL
}
