% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concentration.R
\name{plot_concentration}
\alias{plot_concentration}
\title{Plot concentration}
\usage{
plot_concentration(
  input_xlsx =
    "~/switchdrive/SAPERE/02_raw-data/inhouse/02_sensory/20210329_raw-extract/03_files/20210329_raw-extract.xlsx",
  output = "~/switchdrive/SAPERE/06_figures/figure-raw-extract.pdf"
)
}
\arguments{
\item{input_xlsx}{Input xlsx}

\item{output}{Output}
}
\description{
Plot concentration
}
\examples{
NULL
}
