% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concentration.R
\name{plot_concentration}
\alias{plot_concentration}
\title{Plot concentration}
\usage{
plot_concentration(
  input = system.file("extdata", "concentration_afc.tsv", package = "sapid"),
  output = "~/switchdrive/SAPERE/06_figures/figure-raw-extract.pdf"
)
}
\arguments{
\item{input}{Input}

\item{output}{Output}
}
\description{
Plot concentration
}
\examples{
NULL
}
