% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrices.R
\name{plot_matrices}
\alias{plot_matrices}
\title{Plot matrices}
\usage{
plot_matrices(
  input = system.file("extdata", "profiles.tsv", package = "sapid"),
  output = "~/switchdrive/SAPERE/06_figures/figure_matrices.pdf"
)
}
\arguments{
\item{input}{Input}

\item{output}{Output}
}
\description{
Plot matrices
}
\examples{
NULL

}
