% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_napping.R
\name{plot_napping}
\alias{plot_napping}
\title{Plot napping}
\usage{
plot_napping(
  input_coordinates = "inst/extdata/napping_coordinates.tsv",
  input_descriptors = "inst/extdata/napping_descriptors.tsv",
  sessions = 2
)
}
\arguments{
\item{input_coordinates}{Input coordinates}

\item{input_descriptors}{Input descriptors}

\item{session}{Session}
}
\description{
Plot napping
}
\examples{
NULL
}
