% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_chasselas.R
\name{prepare_chasselas}
\alias{prepare_chasselas}
\title{Prepare chasselas}
\usage{
prepare_chasselas(
  path = data_inhouse_sensory_path,
  path_output = analysis_path_04_output,
  date = DATE,
  product_name = PRODUCTNAME,
  cluster = CLUSTER
)
}
\arguments{
\item{path}{Path}

\item{path_output}{Path output}

\item{date}{Date}

\item{product_name}{Product name}

\item{cluster}{Cluster}
}
\description{
Prepare chasselas
}
\examples{
NULL
}
