% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_chemical_profiles.R
\name{prepare_chemical_profiles}
\alias{prepare_chemical_profiles}
\title{Prepare chemical profiles}
\usage{
prepare_chemical_profiles(
  mzmls_dir = "./data/20210619",
  peaks_dir_extract = "./data/interim/peaks/extract",
  peaks_dir_fractions = "./data/interim/peaks/fractions",
  features_path_extract = "./data/extract_mzmine/extract.csv",
  features_path_fractions = "./data/fractions_mzmine/fractions.csv",
  min_intensity = 1e+05,
  start_extract = 5L,
  end_extract = 7L,
  start_fractions = 12L,
  end_fractions = 65L
)
}
\arguments{
\item{mzmls_dir}{mzML files directory}

\item{peaks_dir_extract}{Peaks dir extract}

\item{peaks_dir_fractions}{Peaks dir fractions}

\item{features_path_extract}{Features path extract}

\item{features_path_fractions}{Features path fractions}

\item{min_intensity}{Minimal intensity}

\item{start_extract}{Start extract}

\item{end_extract}{End extract}

\item{start_fractions}{Start fractions}

\item{end_fractions}{End fractions}
}
\description{
Prepare chemical profiles
}
\examples{
NULL
}
