% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_concentration.R
\name{prepare_concentration}
\alias{prepare_concentration}
\title{Prepare concentration}
\usage{
prepare_concentration(
  input_xlsx =
    "~/switchdrive/SAPERE/02_raw-data/inhouse/02_sensory/20210329_raw-extract/03_files/20210329_raw-extract.xlsx",
  output = "~/git/sapid/inst/extdata/concentration_afc.tsv"
)
}
\arguments{
\item{input_xlsx}{Input xlsx}

\item{output}{Output}
}
\description{
Prepare concentration
}
\examples{
NULL
}
