% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_profiles.R
\name{prepare_profiles}
\alias{prepare_profiles}
\title{Prepare profiles}
\usage{
prepare_profiles(
  input_dir = "~/switchdrive/SAPERE/02_raw-data/inhouse/02_sensory",
  output = "~/git/sapid/inst/extdata/profiles.tsv",
  sessions = seq(1, 7),
  dictionary_generic_path = "inst/extdata/dictionary_generic.tsv",
  dictionary_specific_path = "inst/extdata/dictionary_specific.tsv"
)
}
\arguments{
\item{input_dir}{Input dir}

\item{output}{output}

\item{sessions}{Sessions}

\item{dictionary_generic_path}{Dictionary generic path}

\item{dictionary_specific_path}{Dictionary specific path}
}
\description{
Prepare profiles
}
\examples{
NULL
}
